pyproject.toml
[tool.poetry]
name = "map-tile-client"
version = "0.2.0"
description = ""
authors = ["Nima Rahnemoon <nrahnemoon@gmail.com>"]
readme = "README.md"
packages = [{include = "map_tile_client"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
numpy = "^1.24.2"
requests = "^2.28.2"
pillow = "^9.5.0"
opencv-python = "^4.7.0.72"
scipy = "^1.10.1"
shapely = "^2.0.1"
mapbox-vector-tile = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
ipython = "^8.12.0"
black = "^23.3.0"
flake8 = "^6.0.0"
bumpver = "^2022.1120"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bumpver]
current_version = "0.2.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message  = "Bump version {old_version} -> {new_version}"
commit          = false
tag             = false
push            = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    '^current_version = "{version}"$',
    '^version = "{version}"$',
]
"README.md" = [
    "^VERSION: {version}$",
]
"map_tile_client/__init__.py" = [
    '^__version__ = "{version}"$',
]
